Fix(Google Fonts): Updated tailwind.config.js snippet
Hi! I have been working with the `maizzle new` command, and the snippet of `tailwind.config.js` didn't work. I have set it as you can see in the proposed file changes and now is working for me. Also I have followed this documentation in order to find out this:
https://tailwindcss.com/docs/font-family/

Hope you find it useful 😄